requirements.txt
# LLM & NLP
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-groq>=0.1.0
groq>=0.5.0
openai>=1.10.0
tiktoken>=0.5.2

# Embeddings
sentence-transformers>=2.2.2

# Vector Store
faiss-cpu>=1.8.0

# API Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Dashboard
streamlit>=1.30.0

# HTTP Client
requests>=2.31.0
httpx>=0.27.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Monitoring
prometheus-client>=0.20.0

# Logging
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0

# Linting
ruff>=0.3.0

# Optional: Image Processing (BLIP-2)
# torch>=2.0.0
# transformers>=4.30.0
# Pillow>=10.0.0
